requirements.txt
psycopg2-binary>=2.9
moviepy>=1.0.3
pillow>=10.0.0
numpy>=1.24.0
